Update README.md Readability
The following information was difficult to follow and cluttered together with additional boxes. With the following change additional wording has been modified to provide clear context and repetition is reduced.

Additionally added steps: to provide a clearer path what the user should take to outline them better, in case they are in a hurry and it makes it easier to follow.

Changed directory example to its own codeblock, to make it stand out and clear.